Added Intro / Title ..and
with exception of the array, I have tried to have no hard coded values.
 I believe I have been successful in utilizing user input for every
value needed.  I have two places in the story where user input
influences the direction of the story, user input determines names of
all the characters.  This has been fun.  It’s a dorky story, it may not
be perfect grammatically, but I was more worried about the code and
must move on if I am to get the other two assignments done by next
Wednesday.  Awesome assigment